submissions/BrandView-AI/requirements.txt
agno==1.4.2
boto3==1.38.10
docx2pdf==0.1.8
fastapi==0.115.12
firecrawl-py==2.5.3
fpdf==1.7.2
google-api-python-client==2.168.0
h2==4.2.0
ipykernel==6.29.5
mem0ai==0.1.102
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
pyngrok==7.2.9
python-docx==1.1.2
reportlab==4.4.0
scipy==1.13.1
seaborn==0.13.2
streamlit==1.45.1
tabulate==0.9.0
tavily-python==0.7.2
textblob==0.19.0
torchvision==0.22.0
transformers==4.51.3
twilio==9.6.2
uvicorn==0.34.2
yt-dlp==2025.3.31
